rename(route-folder): 메인 _hooks 폴더 배럴패턴 제거
