+++ src/app/(home)/_components/HomeFilterSection/HomeFilterSection.tsx
@@ -4,7 +4,7 @@ import { useHorizontalDragScroll } from "@/hooks";
 import { Filter } from "@/components";
 import { POST_FILTER_ITEMS, CATEGORY_FILTER_ITEM, CATEGORY, POST_TYPE } from "../HOME_CONST";
 import CategoryFilter from "./_internal/CategoryFilter/CategoryFilter";
-import { useHomeFilterQuery } from "../../_hooks";
+import useHomeFilterQuery from "../../_hooks/useHomeFilterQuery/useHomeFilterQuery";
 
 const HomeFilterSection = ({ isHidden = false }: { isHidden?: boolean }) => {
   const { ref, onMouseDown } = useHorizontalDragScroll();
